python-backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
supabase==2.9.0
python-dotenv==1.0.0
schedule==1.2.1
lxml==5.1.0